Refactor: Refactored sections into their own files instead of having them all in CreateOfferContents
